requirements.txt
Django==1.11.4
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.5.2
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
django-cors-headers==2.1.0
django-filter==1.0.4
django-filters==0.2.1
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
openapi-codec==1.3.2
pip==9.0.1
pytz==2017.2
requests==2.18.4
setuptools==36.3.0
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
wheel==0.29.0
